Move llm-client + llm-facade into src/agent/ (ash-scoped)
Both are ash-internal: LlmClient is the OpenAI-compat HTTP client
ash uses to talk to its providers; llm-facade is what backs
ctx.agent.llm by delegating to ash's llm:invoke handler. Other
backends (claude-code, pi, opencode) bring their own LLM transport
and don't define llm:invoke, so neither file is shared across
backends.

LlmClient stays re-exported from core/index.ts so out-of-tree
consumers that imported it via the agent-sh package root keep
working.
