+++ src/agent/agent-loop.ts
@@ -13,7 +13,7 @@
  */
 import type { EventBus, ShellEvents } from "../core/event-bus.js";
 import type { AgentMode } from "./host-types.js";
-import type { LlmClient } from "../utils/llm-client.js";
+import type { LlmClient } from "./llm-client.js";
 import type { HandlerFunctions } from "../utils/handler-registry.js";
 import { setMaxListeners } from "node:events";
 import * as path from "node:path";
