+++ src/agent/index.ts
@@ -13,8 +13,8 @@ import type { ExtensionContext } from "../shell/host-types.js";
 import type { AgentContext, AgentMode, AgentSurface } from "../agent/host-types.js";
 import type { AppConfig } from "../shell/host-types.js";
 import { AgentLoop } from "./agent-loop.js";
-import { LlmClient } from "../utils/llm-client.js";
-import { createLlmFacade } from "../utils/llm-facade.js";
+import { LlmClient } from "./llm-client.js";
+import { createLlmFacade } from "./llm-facade.js";
 import type { ToolDefinition, ToolSchemaView } from "./types.js";
 import { registerReadOnlyTool, unregisterReadOnlyTool } from "./nuclear-form.js";
 import { resolveProvider, getProviderNames, getSettings, type ResolvedProvider } from "../core/settings.js";
